Minimize the scope of NOINLINE
Previously all the pure tracing functions were marked as NOINLINE.
This prevented GHC from checking userTracingEnabled before calling
into those functions causing a little overhead. Now those pure
tracing functions are split into two functions: the wrappers and
the workers. And only the workers are marked as NOINLINE. The
wrappers do the userTracingEnabled check so if it's False we don't
have to call the NOINLINE'd workers.